pyproject.toml
[tool.poetry]
name = "gradient-sdk-python-example"
version = "0.1.0"
description = ""
authors = ["Gradient developers"]
readme = "README.md"
packages = [{include = "gradient_sdk_python_example"}]

[tool.poetry.dependencies]
gradientai = "^1.2.0"
python = "^3.8.1"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "22.3.0"
mypy = "1.3.0"

[tool.poetry.scripts]
embeddings = "gradient_sdk_python_example.embeddings:main"
fine_tune = "gradient_sdk_python_example.fine_tune:main"
fine_tune_zen = "gradient_sdk_python_example.fine_tune_zen:main"
guidance = "gradient_sdk_python_example.guidance:main"
blocks = "gradient_sdk_python_example.blocks:main"
rag = "gradient_sdk_python_example.rag:main"
guidance2 = "gradient_sdk_python_example.guidance2:main"


[tool.black]
black = "22.3.0"
line-length = 80

[tool.mypy]
check_untyped_defs = true
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_return_any = true
warn_unused_ignores = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
